Add changes for new feature
